+++ Server/Data/Models/Messages/UpsertOrderRequest.cs
@@ -7,7 +7,7 @@ public class UpsertOrderRequest
     public bool Pending { get; set; }
     public EOrderShippingType? ShippingType { get; set; }
     public string? Comment { get; set; }
-    public float? Price { get; set; }
+    public double? Price { get; set; }
     public EOrderTax? Taxes { get; set; }
     public DateTime? DueDate { get; set; }
     public EOrderType? Type { get; set; }
@@ -15,7 +15,7 @@ public class UpsertOrderRequest
     public string? Article { get; set; }
     public string? Color { get; set; }
     public string? Dealer { get; set; }
-    public float? Size { get; set; }
+    public double? Size { get; set; }
     public string? Name { get; set; }
 
     public string? Customer { get; set; }
